added: finished ldap groups service + caching; changed: refactoring
