+++ corex/build.gradle
@@ -80,12 +80,14 @@ dependencies {
     simpleCompile group: 'com.comcast',                           name: 'ip4s-cats_2.12',                         version: '1.1.1'
     shadowCompile group: 'com.fasterxml.jackson.dataformat',      name: 'jackson-dataformat-yaml',                version: '2.8.11'
     shadowCompile group: 'com.fasterxml.jackson.core',            name: 'jackson-databind',                       version: '2.8.11'
+    // todo: remove?
     simpleCompile group: 'com.google.code.gson',                  name: 'gson',                                   version: '2.8.2'
     shadowCompile group: 'com.jayway.jsonpath',                   name: 'json-path',                              version: '2.2.0'
     simpleCompile group: 'com.softwaremill.sttp',                 name: 'async-http-client-backend-cats_2.12',    version: '1.5.8'
     simpleCompile group: 'com.softwaremill.sttp',                 name: 'core_2.12',                              version: '1.5.1'
     simpleCompile group: 'com.unboundid',                         name: 'unboundid-ldapsdk',                      version: '4.0.9'
     simpleCompile group: 'commons-codec',                         name: 'commons-codec',                          version: '1.10'
+    // todo: remove?
     simpleCompile group: 'cz.seznam.euphoria',                    name: 'shaded-guava',                           version: '21.0'
     simpleCompile group: 'eu.timepit',                            name: 'refined_2.12',                           version: '0.9.3'
     simpleCompile group: 'io.circe',                              name: 'circe-core_2.12',                        version: '0.10.1'
@@ -103,6 +105,7 @@ dependencies {
         exclude group: 'org.slf4j', module: 'slf4j-simple'
         exclude group: 'dom4j', module: 'dom4j'
     }
+    simpleCompile group: 'com.github.blemale',                    name: 'scaffeine_2.12',                         version: '2.5.0'
     simpleCompile group: 'org.scala-lang',                        name: 'scala-library',                          version: '2.12.4'
     simpleCompile group: 'org.scala-lang.modules',                name: 'scala-java8-compat_2.12',                version: '0.9.0'
     simpleCompile group: 'io.lemonlabs',                          name: 'scala-uri_2.12',                         version: '1.4.1'
